requirements.txt
opencv-python
pandas
numpy
Pillow
imagehash
openpyxl
tqdm
